Revert "Speed up CLI by lazy loading some imports (#3212)"
This reverts commit 43eaddc8013d484fdfdde91eb07dc6eaf7902601.
