+++ codalab/lib/bundle_cli.py
@@ -51,6 +51,7 @@ from codalab.common import (
     ensure_str,
 )
 from codalab.lib import (
+    bundle_util,
     file_util,
     formatting,
     metadata_util,
@@ -58,6 +59,7 @@ from codalab.lib import (
     spec_util,
     ui_actions,
     worksheet_util,
+    zip_util,
     bundle_fuse,
 )
 from codalab.lib.cli_util import (
@@ -93,8 +95,9 @@ from codalab.lib.completers import (
     WorksheetsCompleter,
 )
 from codalab.lib.bundle_store import MultiDiskBundleStore
+from codalab.lib.interactive_session import InteractiveSession
 from codalab.lib.print_util import FileTransferProgress
-from codalab.worker.un_tar_directory import un_tar_directory
+from codalab.worker.file_util import un_tar_directory
 from codalab.worker.download_util import BundleTarget
 from codalab.worker.bundle_state import State, LinkFormat
 from codalab.rest.worksheet_block_schemas import BlockModes
@@ -1273,8 +1276,6 @@ class BundleCLI(object):
         + EDIT_ARGUMENTS,
     )
     def do_upload_command(self, args):
-        from codalab.lib import zip_util
-
         if args.contents is None and not args.path:
             raise UsageError("Nothing to upload.")
 
@@ -1723,8 +1724,6 @@ class BundleCLI(object):
         targets = self.resolve_key_targets(client, worksheet_uuid, args.target_spec)
 
         if args.interactive:
-            from codalab.lib.interactive_session import InteractiveSession
-
             # Disable cl run --interactive on headless systems
             self._fail_if_headless(args)
 
@@ -2775,8 +2774,6 @@ class BundleCLI(object):
         """
         Use args.bundles to generate a call to bundle_util.mimic_bundles()
         """
-        from codalab.lib import bundle_util
-
         client, worksheet_uuid = self.parse_client_worksheet_uuid(args.worksheet_spec)
         try:
             bundle_uuids = self.target_specs_to_bundle_uuids(client, worksheet_uuid, args.bundles)
