+++ codalab/lib/codalab_manager.py
@@ -48,6 +48,7 @@ from codalab.lib.upload_manager import UploadManager
 from codalab.lib import formatting
 from codalab.model.worker_model import WorkerModel
 
+
 MAIN_BUNDLE_SERVICE = 'https://worksheets.codalab.org'
