+++ codalab/lib/upload_manager.py
@@ -3,7 +3,7 @@ import os
 import shutil
 
 from codalab.common import UsageError
-from codalab.lib import crypt_util, file_util, path_util
+from codalab.lib import crypt_util, file_util, path_util, zip_util
 
 
 class UploadManager(object):
@@ -13,14 +13,11 @@ class UploadManager(object):
     """
 
     def __init__(self, bundle_model, bundle_store):
-        from codalab.lib import zip_util
-
         # exclude these patterns by default
         DEFAULT_EXCLUDE_PATTERNS = ['.DS_Store', '__MACOSX', '^\._.*']
         self._bundle_model = bundle_model
         self._bundle_store = bundle_store
         self._default_exclude_patterns = DEFAULT_EXCLUDE_PATTERNS
-        self.zip_util = zip_util
 
     def upload_to_bundle_store(
         self,
@@ -81,7 +78,7 @@ class UploadManager(object):
                         if unpack and self._can_unpack_file(source_output_path):
                             self._unpack_file(
                                 source_output_path,
-                                self.zip_util.strip_archive_ext(source_output_path),
+                                zip_util.strip_archive_ext(source_output_path),
                                 remove_source=True,
                                 simplify_archive=simplify_archives,
                             )
@@ -92,7 +89,7 @@ class UploadManager(object):
                     if unpack and self._can_unpack_file(source_path):
                         self._unpack_file(
                             source_path,
-                            self.zip_util.strip_archive_ext(source_output_path),
+                            zip_util.strip_archive_ext(source_output_path),
                             remove_source=remove_sources,
                             simplify_archive=simplify_archives,
                         )
@@ -106,11 +103,11 @@ class UploadManager(object):
                             exclude_patterns=exclude_patterns,
                         )
                 elif is_fileobj:
-                    if unpack and self.zip_util.path_is_archive(filename):
+                    if unpack and zip_util.path_is_archive(filename):
                         self._unpack_fileobj(
                             source[0],
                             source[1],
-                            self.zip_util.strip_archive_ext(source_output_path),
+                            zip_util.strip_archive_ext(source_output_path),
                             simplify_archive=simplify_archives,
                         )
                     else:
@@ -139,19 +136,17 @@ class UploadManager(object):
         return is_url, is_local_path, is_fileobj, filename
 
     def _can_unpack_file(self, path):
-        return os.path.isfile(path) and self.zip_util.path_is_archive(path)
+        return os.path.isfile(path) and zip_util.path_is_archive(path)
 
     def _unpack_file(self, source_path, dest_path, remove_source, simplify_archive):
-        self.zip_util.unpack(self.zip_util.get_archive_ext(source_path), source_path, dest_path)
+        zip_util.unpack(zip_util.get_archive_ext(source_path), source_path, dest_path)
         if remove_source:
             path_util.remove(source_path)
         if simplify_archive:
             self._simplify_archive(dest_path)
 
     def _unpack_fileobj(self, source_filename, source_fileobj, dest_path, simplify_archive):
-        self.zip_util.unpack(
-            self.zip_util.get_archive_ext(source_filename), source_fileobj, dest_path
-        )
+        zip_util.unpack(zip_util.get_archive_ext(source_filename), source_fileobj, dest_path)
         if simplify_archive:
             self._simplify_archive(dest_path)
